feat: beauty-talk 페이지 suspense boundary 추가
